Warnings removal checkin.
git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@3054 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpid/ch3/channels/mrail/src/gen2/ring_startup.c
@@ -28,8 +28,8 @@
 #define IBA_PMI_VALLEN  (4096)
 
 struct init_addr_inf {
-    uint16_t    lid;
-    uint32_t    qp_num[2];
+    int    lid;
+    int    qp_num[2];
 };
 
 struct host_addr_inf {
